small test fix 2.0

+++ tests/test_funcnet/TestCouplingAnalysis.py
@@ -30,7 +30,7 @@ def testSymmetrizeByAbsmax():
     tdata = create_test_data()
     n_index, n_times = tdata.shape
     coup_ana = CouplingAnalysis(tdata)
-    similarity_matrix = np.random.rand(n_index, n_times)
+    similarity_matrix = np.random.rand(n_index, n_times).astype('float32')
     lag_matrix = np.random.rand(n_index, n_times).astype(np.int)
     sm_new = coup_ana.symmetrize_by_absmax(similarity_matrix, lag_matrix)[0]
     for i in range(n_index):
